Adding a main application concept.
Now that we are using blueprints, we should have a main `application` object.
This is living under `labmanager/application.py`. The goal of this app file
is to have a main object representing the whole application. Specially the
public parts.

Instead of always importing everything, we can choose only some functions where
needed.