Fix show-object (admin view) functionality
To have this functionality we will have to ignore the fact that some
models are not being used in the view/admin/main.py file, or find a
way to let sys.modules know about classes that are not loaded in the
file.

Using the .has_access method on the laboratories.