Added support to recreate a development VM using Vagrant
The support is written to install a default Ubuntu 12.04 LTS instance,
the provisioning is done via shell commands which are simpler to start
than a full Chef or Puppet environment, and that additionally can be
moved to a Dockerfile in the future.

The provision is divided in two parts, first one consists in the default
installation of system packages (root). And the second one consists in
the concrete creation of the python virtual environment, and its
dependencies.